turn off the LED sooner when out of data
It was checking to see if there was any more data, and if the next
byte wants the LED on it would leave it on, but for the datalink
protocol there aren't any two adjacent on bytes, turn it off before
checking for more data that keeps the duration more consistent.
At some point the watch stopped receiving data, need to figure that
out.